Also move param code into its own file, it was getting unwieldy
Created using spr 1.3.6-beta.1
