+++ dropshot_endpoint/src/endpoint.rs
@@ -2,8 +2,6 @@
 
 //! Support for HTTP `#[endpoint]` macros.
 
-use std::fmt;
-
 use proc_macro2::TokenStream;
 use quote::format_ident;
 use quote::quote;
@@ -13,11 +11,13 @@ use serde::Deserialize;
 use serde_tokenstream::from_tokenstream;
 use serde_tokenstream::Error;
 use syn::spanned::Spanned;
-use syn::visit::Visit;
 
 use crate::doc::ExtractedDoc;
 use crate::error_store::ErrorSink;
 use crate::error_store::ErrorStore;
+use crate::params::validate_param_ty;
+use crate::params::ParamTyKind;
+use crate::params::RqctxTy;
 use crate::syn_parsing::ItemFnForSignature;
 use crate::util::get_crate;
 use crate::util::ValidContentType;
@@ -499,204 +499,6 @@ impl<'ast> EndpointParams<'ast> {
                     future_endpoint_must_be_send(#name(#(#arg_names),*));
                 }
             };
-        }
-    }
-}
-
-/// Perform syntactic validation for an argument or return type.
-///
-/// This returns the input type if it is valid.
-fn validate_param_ty<'ast>(
-    ty: &'ast syn::Type,
-    kind: ParamTyKind,
-    name_str: &str,
-    errors: &ErrorSink<'_, Error>,
-) -> Option<&'ast syn::Type> {
-    // Types can be arbitrarily nested, so to keep these checks simple we use
-    // the visitor pattern.
-
-    let errors = errors.new();
-
-    // This just needs a second 'store lifetime because the one inside ErrorSink
-    // is invariant. Everything else is covariant and can share lifetimes.
-    struct Visitor<'store, 'ast> {
-        kind: ParamTyKind,
-        name_str: &'ast str,
-        errors: &'ast ErrorSink<'store, Error>,
-    }
-
-    impl<'store, 'ast> Visit<'ast> for Visitor<'store, 'ast> {
-        fn visit_bound_lifetimes(&mut self, i: &'ast syn::BoundLifetimes) {
-            let name_str = self.name_str;
-            let kind = self.kind;
-            self.errors.push(Error::new_spanned(
-                i,
-                format!(
-                    "endpoint `{name_str}` must not have lifetime bounds \
-                     in {kind}",
-                ),
-            ));
-        }
-
-        fn visit_lifetime(&mut self, i: &'ast syn::Lifetime) {
-            let name_str = self.name_str;
-            let kind = self.kind;
-            if i.ident != "static" {
-                self.errors.push(Error::new_spanned(
-                    i,
-                    format!(
-                        "endpoint `{name_str}` must not have lifetime parameters \
-                         in {kind}",
-                    ),
-                ));
-            }
-        }
-
-        fn visit_ident(&mut self, i: &'ast syn::Ident) {
-            if i == "Self" {
-                let name_str = self.name_str;
-                let kind = self.kind;
-                self.errors.push(Error::new_spanned(
-                    i,
-                    format!(
-                        "endpoint `{name_str}` must not have `Self` in {kind}",
-                    ),
-                ));
-            }
-        }
-
-        fn visit_type_impl_trait(&mut self, i: &'ast syn::TypeImplTrait) {
-            let name_str = self.name_str;
-            let kind = self.kind;
-            self.errors.push(Error::new_spanned(
-                i,
-                format!(
-                    "endpoint `{name_str}` must not have impl Trait in {kind}",
-                ),
-            ));
-        }
-    }
-
-    let mut visitor = Visitor { kind, name_str, errors: &errors };
-    visitor.visit_type(ty);
-
-    // Don't return the type if there were errors.
-    (!errors.has_errors()).then(|| ty)
-}
-
-/// A representation of the RequestContext type.
-#[derive(Clone, Eq, PartialEq)]
-enum RqctxTy<'ast> {
-    /// This is a function-based macro, with the payload being the full type.
-    Function(&'ast syn::Type),
-}
-
-impl<'ast> RqctxTy<'ast> {
-    /// Ensures that the type parameter for RequestContext is valid.
-    fn new(
-        name_str: &str,
-        ty: &'ast syn::Type,
-        errors: &ErrorSink<'_, Error>,
-    ) -> Option<Self> {
-        let param = match extract_rqctx_param(ty) {
-            Ok(Some(ty)) => ty,
-            Ok(None) => {
-                return Some(Self::Function(ty));
-            }
-            Err(_) => {
-                // Can't do any further validation on the type.
-                errors.push(Error::new_spanned(
-                    ty,
-                    format!(
-                        "endpoint `{name_str}` must accept a \
-                        RequestContext<T> as its first argument",
-                    ),
-                ));
-                return None;
-            }
-        };
-
-        // Now validate the inner parameter.
-        validate_param_ty(param, ParamTyKind::RequestContext, name_str, errors)
-            .map(|_| Self::Function(ty))
-    }
-
-    fn as_type(&self) -> &'ast syn::Type {
-        match self {
-            RqctxTy::Function(ty) => ty,
-        }
-    }
-}
-
-impl<'ast> fmt::Debug for RqctxTy<'ast> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            RqctxTy::Function(ty) => write!(f, "Function({})", quote! { #ty }),
-        }
-    }
-}
-
-/// Extracts and ensures that the type parameter for RequestContext is valid.
-fn extract_rqctx_param<'ast>(
-    ty: &'ast syn::Type,
-) -> Result<Option<&'ast syn::Type>, RqctxTyError> {
-    let syn::Type::Path(p) = &*ty else {
-        return Err(RqctxTyError::NotTypePath);
-    };
-
-    // Inspect the last path segment.
-    let Some(last_segment) = p.path.segments.last() else {
-        return Err(RqctxTyError::NoPathSegments);
-    };
-
-    // It must either not have type arguments at all, or if so then exactly one
-    // argument.
-    let a = match &last_segment.arguments {
-        syn::PathArguments::None => {
-            // This is all right -- hopefully a type alias.
-            return Ok(None);
-        }
-        syn::PathArguments::AngleBracketed(a) => a,
-        syn::PathArguments::Parenthesized(_) => {
-            // This isn't really possible in this position?
-            return Err(RqctxTyError::ArgsNotAngleBracketed);
-        }
-    };
-
-    if a.args.len() != 1 {
-        return Err(RqctxTyError::IncorrectTypeArgCount(a.args.len()));
-    }
-
-    // The argument must be a type.
-    let syn::GenericArgument::Type(tp) = a.args.first().unwrap() else {
-        return Err(RqctxTyError::ArgNotType);
-    };
-
-    Ok(Some(tp))
-}
-
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-enum RqctxTyError {
-    NotTypePath,
-    NoPathSegments,
-    ArgsNotAngleBracketed,
-    IncorrectTypeArgCount(usize),
-    ArgNotType,
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-enum ParamTyKind {
-    RequestContext,
-    Extractor,
-    Return,
-}
-
-impl fmt::Display for ParamTyKind {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            ParamTyKind::RequestContext => write!(f, "RequestContext"),
-            ParamTyKind::Extractor => write!(f, "extractor"),
-            ParamTyKind::Return => write!(f, "return type"),
         }
     }
 }
