+++ dropshot_endpoint/src/lib.rs
@@ -13,6 +13,7 @@ mod channel;
 mod doc;
 mod endpoint;
 mod error_store;
+mod params;
 mod syn_parsing;
 mod util;
